+++ lib/mongoid/full_text_search/config.rb
@@ -1,0 +1,11 @@
+module Mongoid
+  module FullTextSearch
+    module Config
+      extend ActiveSupport::Concern
+
+      included do
+        cattr_accessor :mongoid_fulltext_config
+      end
+    end
+  end
+end
